@@ -1,7 +1,7 @@
 import { ValidationValue } from './validation-value';
 import { ValidationCombinationResult } from './validation-combination-result';
 
-export interface ValidationCombination { //TODO rename api objects to responses'
+export interface ValidationCombination {
 
   id: number;
   validationValue: ValidationValue,
